+++ src/rex.setup/src/rex/setup/__init__.py
@@ -18,7 +18,7 @@ from .generate import Generate, watch
 from .download import GenerateDownload
 from .webpack import GenerateWebpack
 from .doc import GenerateDoc
-from .commonjs import node, npm
+from .commonjs import node, npm, install_commonjs, develop_commonjs
 import os