+++ src/rex.setup/src/rex/setup/commonjs/polyfills/object-assign.js
@@ -1,3 +1,0 @@
-if (Object.assign === undefined) {
-  Object.assign = require('object-assign');
-}